+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
 ## [Unreleased]
+###
+- Added mongo indexes to optimize performance
 
 ## [2.1.0] - 2026-04-07
 ### Added